Made sure that tests and Swig still work even if Petsc is disabled.
