+++ CMakeLists.txt
@@ -6,7 +6,11 @@ set (CDMATH_VERSION_MAJOR 0)
 set (CDMATH_VERSION_MINOR 2)
 
 # Project options
-option(CDMATH_WITH_PETSC "Compile CDMATH with PETSc linking" OFF)
+option (CDMATH_WITH_PETSC "Compile CDMATH with PETSc linking." OFF)
+option (CDMATH_WITH_DOCUMENTATION "Generate documentation for CDMATH use." OFF)
+option (CDMATH_WITH_PYTHON "Compile Python interface for CDMATH." OFF)
+option (CDMATH_WITH_PACKAGE "Generate RPM, Debian and tarball packages." OFF)
+option (CDMATH_WITH_TESTS "Compile unit testing." OFF)
 
 # Base directories
 set (BASE_DIR ${CDMATH_SOURCE_DIR}/base)
@@ -41,24 +45,24 @@ if (CDMATH_WITH_PETSC)
 endif()
 
 # Swig interface
-if (CMAKE_CDMATH_SWIG)
+if (CDMATH_WITH_SWIG)
    find_package (PYTHON REQUIRED)
    find_package (SWIG REQUIRED)
    set (CPACK_DEBIAN_PACKAGE_DEPENDS "${CPACK_DEBIAN_PACKAGE_DEPENDS}, python-dev, python-numpy, swig")
    set (CPACK_RPM_PACKAGE_REQUIRES "${CPACK_RPM_PACKAGE_REQUIRES}, python-devel, numpy, swig")
    add_subdirectory (${CDMATH_SWIG_DIR})
-endif (CMAKE_CDMATH_SWIG)
+endif (CDMATH_WITH_SWIG)
 
 add_subdirectory (${PRE_REQUIS_DIR})
 add_subdirectory (${BASE_DIR})
 add_subdirectory (${MESH_DIR})
 
-if(CDMATH_WITH_PETSC)
+if (CDMATH_WITH_PETSC)
     add_subdirectory (${LINEARSOLVER_DIR})
-endif()
+endif (CDMATH_WITH_PETSC)
 
 # Documentation
-if (CMAKE_CDMATH_DOCUMENTATION)
+if (CDMATH_WITH_DOCUMENTATION)
   find_package (Doxygen)
   set (CPACK_DEBIAN_PACKAGE_DEPENDS "${CPACK_DEBIAN_PACKAGE_DEPENDS}, doxygen, graphviz, mscgen")
   set (CPACK_RPM_PACKAGE_REQUIRES "${CPACK_RPM_PACKAGE_REQUIRES}, doxygen, graphviz, mscgen")
@@ -69,21 +73,21 @@ if (CMAKE_CDMATH_DOCUMENTATION)
   add_custom_target (doc ALL COMMAND ${DOXYGEN_EXECUTABLE} ${PROJECT_BINARY_DIR}/Doxyfile
   							  SOURCES ${PROJECT_BINARY_DIR}/Doxyfile)
   install (DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/doc/html DESTINATION share/doc/cdmath)
-endif (CMAKE_CDMATH_DOCUMENTATION)
+endif (CDMATH_WITH_DOCUMENTATION)
  
 # Tests
-if (CMAKE_CDMATH_TESTS)
+if (CDMATH_WITH_TESTS)
   set (CMAKE_EXE_LINKER_FLAGS "-pg -fprofile-arcs -ftest-coverage -lgcov")
   enable_testing ()
   find_package (CPPUNIT REQUIRED)
   add_subdirectory (${TESTS_DIR})
-endif (CMAKE_CDMATH_TESTS)
+endif (CDMATH_WITH_TESTS)
 
 # Examples
 set (examples_dir ${CDMATH_SOURCE_DIR}/examples)
 add_subdirectory (${examples_dir})
 
 # Packaging
-if (CMAKE_CDMATH_PACKAGE)
+if (CDMATH_WITH_PACKAGE)
   include (CPackLists.txt)
-endif (CMAKE_CDMATH_PACKAGE)
+endif (CDMATH_WITH_PACKAGE)
