+++ swig/CMakeLists.txt
@@ -4,28 +4,50 @@ SET_SOURCE_FILES_PROPERTIES(cdmath.i PROPERTIES CPLUSPLUS ON)
 SET_SOURCE_FILES_PROPERTIES(cdmath.i PROPERTIES C OFF)
 SET_SOURCE_FILES_PROPERTIES(cdmath.i PROPERTIES SWIG_DEFINITIONS "-shadow")
 
-INCLUDE_DIRECTORIES(
-  ${PYTHON_INCLUDE_DIRS}
-  ${HDF5_INCLUDE_DIRS}
-  ${MED3_INCLUDE_DIRS}
-  ${MED_COUPLING_DIR}
-  ${MED_COUPLING_SWIG_DIR}
-  ${BASE_DIR}/inc
-  ${MESH_DIR}/inc
-  ${INTERP_KERNEL_DIR}
-  ${INTERP_KERNEL_DIR}/Bases
-  ${INTERP_KERNEL_DIR}/Geometric2D
-  ${INTERP_KERNEL_DIR}/ExprEval
-  ${INTERP_KERNEL_DIR}/GaussPoints
-  ${MED_LOADER_DIR}
-  ${PETSC_INCLUDES}
-  ${LINEARSOLVER_DIR}/inc
-)
+if (CDMATH_WITH_PETSC)
+    INCLUDE_DIRECTORIES(
+      ${PYTHON_INCLUDE_DIRS}
+      ${HDF5_INCLUDE_DIRS}
+      ${MED3_INCLUDE_DIRS}
+      ${MED_COUPLING_DIR}
+      ${MED_COUPLING_SWIG_DIR}
+      ${BASE_DIR}/inc
+      ${MESH_DIR}/inc
+      ${INTERP_KERNEL_DIR}
+      ${INTERP_KERNEL_DIR}/Bases
+      ${INTERP_KERNEL_DIR}/Geometric2D
+      ${INTERP_KERNEL_DIR}/ExprEval
+      ${INTERP_KERNEL_DIR}/GaussPoints
+      ${MED_LOADER_DIR}
+      ${PETSC_INCLUDES}
+      ${LINEARSOLVER_DIR}/inc
+    )
+else (CDMATH_WITH_PETSC)
+    INCLUDE_DIRECTORIES(
+      ${PYTHON_INCLUDE_DIRS}
+      ${HDF5_INCLUDE_DIRS}
+      ${MED3_INCLUDE_DIRS}
+      ${MED_COUPLING_DIR}
+      ${MED_COUPLING_SWIG_DIR}
+      ${BASE_DIR}/inc
+      ${MESH_DIR}/inc
+      ${INTERP_KERNEL_DIR}
+      ${INTERP_KERNEL_DIR}/Bases
+      ${INTERP_KERNEL_DIR}/Geometric2D
+      ${INTERP_KERNEL_DIR}/ExprEval
+      ${INTERP_KERNEL_DIR}/GaussPoints
+      ${MED_LOADER_DIR}
+    )
+endif (CDMATH_WITH_PETSC)
 
 
 SWIG_ADD_MODULE(cdmath python cdmath.i)
 
-SWIG_LINK_LIBRARIES(cdmath ${PYTHON_LIBS} base mesh linearsolver medcoupling ${MED3_LIBS} medloader ${PETSC_LIBRARIES})
+if (CDMATH_WITH_PETSC)
+    SWIG_LINK_LIBRARIES(cdmath ${PYTHON_LIBS} base mesh linearsolver medcoupling ${MED3_LIBS} medloader ${PETSC_LIBRARIES})
+else (CDMATH_WITH_PETSC)
+    SWIG_LINK_LIBRARIES(cdmath ${PYTHON_LIBS} base mesh medcoupling ${MED3_LIBS} medloader)
+endif (CDMATH_WITH_PETSC)
 
 INSTALL(TARGETS _cdmath DESTINATION lib/cdmath)
 INSTALL(FILES ${CMAKE_CURRENT_BINARY_DIR}/cdmath.py DESTINATION bin/cdmath)
