+++ tests/CMakeLists.txt
@@ -4,19 +4,30 @@ CMAKE_MINIMUM_REQUIRED(VERSION 2.8)
 
 ADD_SUBDIRECTORY(cdmath)
 IF (CMAKE_CDMATH_SWIG)
-ADD_SUBDIRECTORY(swig)
+    ADD_SUBDIRECTORY(swig)
 ENDIF (CMAKE_CDMATH_SWIG)
 
-INCLUDE_DIRECTORIES(
-  ${BASE_DIR}/inc 
-  ${MESH_DIR}/inc 
-  ${LINEARSOLVER_DIR}/inc
-  ${MED_COUPLING_DIR}
-  ${PRE_REQUIS_DIR}/INTERP_KERNEL
-  ${PRE_REQUIS_DIR}/INTERP_KERNEL/Bases
-  ${PETSC_INCLUDES} 
-  cdmath 
-  )
+if (CDMATH_WITH_PETSC)
+    INCLUDE_DIRECTORIES(
+      ${BASE_DIR}/inc 
+      ${MESH_DIR}/inc 
+      ${LINEARSOLVER_DIR}/inc
+      ${MED_COUPLING_DIR}
+      ${PRE_REQUIS_DIR}/INTERP_KERNEL
+      ${PRE_REQUIS_DIR}/INTERP_KERNEL/Bases
+      ${PETSC_INCLUDES} 
+      cdmath 
+      )
+else (CDMATH_WITH_PETSC)
+    INCLUDE_DIRECTORIES(
+      ${BASE_DIR}/inc 
+      ${MESH_DIR}/inc
+      ${MED_COUPLING_DIR}
+      ${PRE_REQUIS_DIR}/INTERP_KERNEL
+      ${PRE_REQUIS_DIR}/INTERP_KERNEL/Bases 
+      cdmath 
+      )
+endif (CDMATH_WITH_PETSC)
 
 SET(MAIN_SOURCES
   MainTests.cxx
@@ -24,6 +35,10 @@ SET(MAIN_SOURCES
 
 ADD_EXECUTABLE(TestCDMATH ${MAIN_SOURCES})
 SET_TARGET_PROPERTIES(TestCDMATH PROPERTIES COMPILE_FLAGS "")
-TARGET_LINK_LIBRARIES(TestCDMATH TestMesh base mesh linearsolver medcoupling ${HDF5_LIBRARIES} ${MED3_LIBS} medloader ${CPPUNIT_LIBS} ${PETSC_LIBRARIES})
+if (CDMATH_WITH_PETSC)
+    TARGET_LINK_LIBRARIES(TestCDMATH TestMesh base mesh linearsolver medcoupling ${HDF5_LIBRARIES} ${MED3_LIBS} medloader ${CPPUNIT_LIBS} ${PETSC_LIBRARIES})
+else (CDMATH_WITH_PETSC)
+    TARGET_LINK_LIBRARIES(TestCDMATH TestMesh base mesh medcoupling ${HDF5_LIBRARIES} ${MED3_LIBS} medloader ${CPPUNIT_LIBS})
+endif (CDMATH_WITH_PETSC)
 ADD_TEST(TestCDMATH TestCDMATH)
 INSTALL(TARGETS TestCDMATH DESTINATION bin/tests)
