+++ tests/MainTests.cxx
@@ -2,7 +2,7 @@
  * tests_mesh.cxx
  *
  *  Created on: 24 janv. 2012
- *      Authors: CDMAT
+ *      Authors: CDMATH
  */
 
 #include <cppunit/extensions/HelperMacros.h>
@@ -25,7 +25,9 @@
 #include "MeshTests.hxx"
 #include "DoubleTabTests.hxx"
 #include "IntTabTests.hxx"
-#include "LinearSolverTests.hxx"
+#ifdef _CDMATH_WITH_PETSC
+    #include "LinearSolverTests.hxx"
+#endif
 
 #include <iostream>
 
@@ -34,34 +36,36 @@ using namespace std;
 
 int main( int argc, char* argv[] )
 {
-   // Create the event manager and test controller
-   CppUnit::TestResult controller;
-   // Add a listener that colllects test result
-   CppUnit::TestResultCollector result;
-   controller.addListener( &result );
-   // Add a listener that print dots as test run.
-   CppUnit::BriefTestProgressListener progress;
-   controller.addListener( &progress );
-   // Add the top suite to the test runner
-   CppUnit::TextUi::TestRunner runner;
-   runner.addTest( IntTabTests::suite() );
-   runner.addTest( DoubleTabTests::suite() );
-   runner.addTest( MatrixTests::suite() );
-   runner.addTest( SparseMatrixTests::suite() );
-   runner.addTest( PointTests::suite() );
-   runner.addTest( NodeTests::suite() );
-   runner.addTest( CellTests::suite() );
-   runner.addTest( FaceTests::suite() );
-   runner.addTest( FieldTests::suite() );
-   runner.addTest( MeshTests::suite() );
-   runner.addTest( VectorTests::suite() );
-   runner.addTest( LinearSolverTests::suite() );
+    // Create the event manager and test controller
+    CppUnit::TestResult controller;
+    // Add a listener that colllects test result
+    CppUnit::TestResultCollector result;
+    controller.addListener( &result );
+    // Add a listener that print dots as test run.
+    CppUnit::BriefTestProgressListener progress;
+    controller.addListener( &progress );
+    // Add the top suite to the test runner
+    CppUnit::TextUi::TestRunner runner;
+    runner.addTest( IntTabTests::suite() );
+    runner.addTest( DoubleTabTests::suite() );
+    runner.addTest( MatrixTests::suite() );
+    runner.addTest( SparseMatrixTests::suite() );
+    runner.addTest( PointTests::suite() );
+    runner.addTest( NodeTests::suite() );
+    runner.addTest( CellTests::suite() );
+    runner.addTest( FaceTests::suite() );
+    runner.addTest( FieldTests::suite() );
+    runner.addTest( MeshTests::suite() );
+    runner.addTest( VectorTests::suite() );
+    #ifdef _CDMATH_WITH_PETSC
+        runner.addTest( LinearSolverTests::suite() );
+    #endif
 
-   runner.run( controller );
-   CppUnit::CompilerOutputter outputter( &result, CppUnit::stdCOut() );
-   outputter.write();
-   ofstream xmlFileOut("cppUnitResults.xml");
-   CppUnit::XmlOutputter xmlOut(&result, xmlFileOut);
-   xmlOut.write();
-   return result.wasSuccessful() ? 0 : 1;
+    runner.run( controller );
+    CppUnit::CompilerOutputter outputter( &result, CppUnit::stdCOut() );
+    outputter.write();
+    ofstream xmlFileOut("cppUnitResults.xml");
+    CppUnit::XmlOutputter xmlOut(&result, xmlFileOut);
+    xmlOut.write();
+    return result.wasSuccessful() ? 0 : 1;
 }
