+++ tests/swig/CMakeLists.txt
@@ -1,6 +1,7 @@
 
 SET(TestMesh_SOURCES
   mainTests.py
+  linearSolverTests.py
   )
 
 SET(MESH_MED
@@ -9,12 +10,17 @@ SET(MESH_MED
 
 file(COPY ${MESH_MED} DESTINATION tests/swig) 
 ADD_TEST(TestsCDMATHSwig ${PYTHON_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/mainTests.py)
+ADD_TEST(TestsLinearSolverSwig ${PYTHON_EXECUTABLE} ${CMAKE_CURRENT_SOURCE_DIR}/linearSolverTests.py)
 
 # Some users have to set environment variables manually before launching tests.
-SET_PROPERTY(TEST TestsCDMATHSwig PROPERTY ENVIRONMENT "LD_LIBRARY_PATH=${CMAKE_BINARY_DIR}/lib:${PETSC_DIR}/lib:$ENV{LD_LIBRARY_PATH}")
+SET_PROPERTY(TEST TestsCDMATHSwig PROPERTY ENVIRONMENT "LD_LIBRARY_PATH=${CMAKE_BINARY_DIR}/lib:$ENV{LD_LIBRARY_PATH}")
 SET_PROPERTY(TEST TestsCDMATHSwig APPEND PROPERTY ENVIRONMENT "PYTHONPATH=${CMAKE_BINARY_DIR}/lib/cdmath:${CMAKE_BINARY_DIR}/bin/cdmath")
+if (CDMATH_WITH_PETSC)
+    SET_PROPERTY(TEST TestsLinearSolverSwig PROPERTY ENVIRONMENT "LD_LIBRARY_PATH=${CMAKE_BINARY_DIR}/lib:${PETSC_DIR}/lib:$ENV{LD_LIBRARY_PATH}")
+endif (CDMATH_WITH_PETSC)
 #MESSAGE("LD_LIB : $ENV{LD_LIBRARY_PATH}")
 #MESSAGE("PYPATH : $ENV{PYTHONPATH}")
 
 INSTALL(FILES mesh.med DESTINATION bin/tests)
 INSTALL(FILES mainTests.py DESTINATION bin/tests)
+INSTALL(FILES linearSolver.py DESTINATION bin/tests)
